+++ nc_py_api/files/files.py
@@ -333,7 +333,7 @@ class FilesAPI:
         headers = Headers({"Destination": dest}, encoding="utf-8")
         response = self._session.adapter_dav.request(
             "MOVE",
-            f"/versions/{self._session.user}/{file_object.user_path}",
+            quote(f"/versions/{self._session.user}/{file_object.user_path}"),
             headers=headers,
         )
         check_error(response, f"restore_version: user={self._session.user}, src={file_object.user_path}")
@@ -813,7 +813,7 @@ class AsyncFilesAPI:
         headers = Headers({"Destination": dest}, encoding="utf-8")
         response = await self._session.adapter_dav.request(
             "MOVE",
-            f"/versions/{await self._session.user}/{file_object.user_path}",
+            quote(f"/versions/{await self._session.user}/{file_object.user_path}"),
             headers=headers,
         )
         check_error(response, f"restore_version: user={await self._session.user}, src={file_object.user_path}")