+++ tests/_app_security_checks.py
@@ -34,25 +34,12 @@ if __name__ == "__main__":
     result = httpx.put(request_url, headers=headers)
     assert result.status_code == 200
     # Invalid EX-APP-ID
-    old_app_name = headers["EX-APP-ID"]
+    old_app_name = headers.get("EX-APP-ID")
     headers["EX-APP-ID"] = "unknown_app"
     sign_request(headers)
     result = httpx.put(request_url, headers=headers)
     assert result.status_code == 401
     headers["EX-APP-ID"] = old_app_name
-    sign_request(headers)
-    result = httpx.put(request_url, headers=headers)
-    assert result.status_code == 200
-    # Invalid EX-APP-VERSION
-    sign_request(headers)
-    result = httpx.put(request_url, headers=headers)
-    assert result.status_code == 200
-    old_version = headers["EX-APP-VERSION"]
-    headers["EX-APP-VERSION"] = "999.0.0"
-    sign_request(headers)
-    result = httpx.put(request_url, headers=headers)
-    assert result.status_code == 401
-    headers["EX-APP-VERSION"] = old_version
     sign_request(headers)
     result = httpx.put(request_url, headers=headers)
     assert result.status_code == 200